divide algorithms into more files
