+++ src/ui/AutomatonWindow.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useContext } from "react";
 import { CoreContext } from "../core/CoreContext";
 import { Kind } from "../core/core";
-import { TestingAlgorithm } from "../engine/algorithm";
+import { TestingAlgorithm } from "../engine/algorithm/algorithm";
 import { PRIMARY_CYTOSCAPE_ID, SECONDARY_CYTOSCAPE_ID } from "../constants";
 
 export default function AutomatonWindow({ primary, cols }: { primary: boolean, cols: number }) {
